services/exchange_monitor/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
psycopg2-binary==2.9.9
ccxt==4.4.27
requests==2.32.3
python-multipart==0.0.12
